Add tracking of last move direction in Square model
This update includes a new attribute 'lastMoveDir' in the Square case class and logic to set its value in the moveUnitTo function. Further, updates were made in the UiLayer.scala file to use this attribute and determine the drawing direction of the unit. This will ensure that unit images are drawn accurately according to their movement direction.