Refactor code and add game download feature
Refactored the code to improve the clarity of the 'refreshComLayer' function in the Game.scala file. Introduced helper functions for checking terrain and opponent conditions. In the UiController.scala file, we added support for downloading the current game state by pressing the 'D' key. This generates a .txt file that represents the current terrain layout in the game.