Update line and shape drawing methods on UiLayer
The lineWidth default value in the drawLine method has been reduced to 1. The rect method that drew a rectangle in the UiLayer has been replaced by an ellipse method, which draws an ellipse instead. To ensure a more accurate type conversion, the toLinear method in the Point class now does a direct conversion to Int instead of rounding off first.