Update cursor style based on hover square state
The commit allows the cursor style to be updated based on whether or not the hovered square can be moved in the game. It makes sure to set the cursor to 'pointer' when movement is possible and to 'default' when movement is not allowed. This helps with user interaction and provides a better game experience.