Add lobby message input and send feature
Changes have been made to implement a lobby message input and send feature. These changes include updating the UI to include an input field and a send button for the lobby, and handling the "Enter" key press event and the "click" event on the send button to send the message. In addition, updates were made to the matching keycodes for the next turn and game download functionalities. Minor updates have been made to the server caching headers as well.